chore(peg/1): Clean up how failure is handled
This change started on the hope to understand better the failure state
for predicates. The boolean flag `m->failed` is now mostly to signal
failure to the outside of the VM. Although it's still going to cause
the VM to quit if the fail state is set to true and the VM isn't
backtracking. That made it easier to implement fail twice but I still
don't understand why both cases for not (with & without fail twice)
aren't backtracking the input cursor `m->i`.