fix(peg/1): Separate loading from reading bytecode & new fail state
The fail state is being hard to get right, but here's a new
attempt. Now the failure isn't reported by m->fail anymore (that
should be removed if this experiment goes well). The result of mEval
is now what signals success or failure. If there's a match, the output
will be how far in the string the match got. Only NULL means failure.

Separating the reading of the bytecode from its execution improves on
two points:

 1. Now all the address offsets can ignore the size of the pieces of
    an instruction and just jump X instructions from where PC is. So
    much less error prone. Some tests had wrong offsets and it wasn't
    ever going to work.

 2. With this separation, the function mRead already takes care of
    reading the sign flag within the argument part of te instruction
    so now we properly support negative jumps! (That's a requirement
    for the Star operation to work).

The one drawback is that now we'll have an upfront cost of reading the
instructions during the VM boot.

This commit also added the weird header that I've been using to debug
things. Must definitely be improved but it's working for now.

There's one test broken but I already had so much going on that it'll
need to be fixed in the next commit.