feat(vm): Extract parse tree from captured input matches
This feature isn't really finished since the compiler support isn't in
a very good shape, so there aren't extensive tests for it yet.

Two new instructions were added to the parsing machine:
OP_CAP_{OPEN,CLOSE}. Although the OP_SET instruction isn't implemented
yet, its opcode was added so it follows OP_SPAN.

The fields `captures' and `cap' were added to the parsing machine's
state. `captures' is a stack of CaptureEntry objects and `cap' points
to the top of that stack. New entries get pushed to the stack during
the machine execution. They store the id of the matching rule and the
position within the input. After the matching is done, the captures
can be extracted from the stack and factored into a parse tree. This
functionality is implemented by the function `mExtract'.

The values within the output parse tree and the parse tree itself are
built with the new module `value' which currently implements cons
cells, atoms and Nil and add some basic functionality to create new
objects.

There's currently no memory management to objects being created by the
extraction.