feat(cc.py): Optionally compile capture instructions
This feature isn't really perfect. The capture sequence generated for
sequences and for the star operator seem to be off. This is getting in
because the VM support looks good enough for now and the work in the
compiler isn't complete but is in a good stopping point. Captures are
turned off by default.