feat(cc.py): Build string table & add it at top of Bytecode file
All the strings needed to display the names of matches during run-time
are now captured by the compiler within a table that gets dumped at
the header of the binary file.

The identifiers of the rules are added to the string table as well as
the class or literal that the input can be matched against. During
this task, the capture system was changed in the following way:

 1. Don't generate captures for sequences. Maybe will have to do that
    for sequence within sequence. Not sure yet.

 2. Simplify captures for classes for now. Don't try to generate
    anything special, just let the terminals capture values for
    themselves.

These changes made the number of parentheses generated decrease a lil
bit which made it all easier to debug and the output should obviously
be as simple as possible as long as it's complete.