fix(vm): Replace current extract with top down & use a stack
After all the wrestling with this piece of code, it's finally looking
good. Instead of building the tree bottom up, it's now being done top
down with a stack.

  Given the grammar:
    Assign <- Name
    Name <- [a-z]+

  And Input:
    id

  Its Captures should go like:
    push Assign    | Assign
    push Name      | Assign Name
    push (a-z . i) | Assign Name (a-z . i)
    push (a-z . d) | Assign Name (a-z . i) (a-z . d)
    reduce         | Assign (Name (a-z . i) (a-z . d))
    reduce         | (Assign (Name (a-z . i) (a-z . d)))

It basically pushes the atoms of the names of non terminals as soon as
they're found. Non terminals are pushed as cons cells initially (with
the key that matched them). When we're handling non terminals, we pop
all the items in the stack that are cons cells and cons it with the
non terminal key.