feat(cc.py): Introduce the Quiet(;) operator to PEGe
The current compiler currently defaults to capturing every possible
node in the parsing tree. The Quiet operator was introduced to allow
skipping capturing nodes marked with it.

The Quiet operator is only effective on Expressions but can be used
for both terminals and non-terminals. The only expression that can't
be skipped (syntactically and in practice) is the expression of the
Start rule.

The Quiet operator *IS NOT RECURSIVE* which means that if an
identifier of an expression is silenced with the Quiet operator,
values didn't explicitly silenced will be captured. That's very useful
for getting rid of intermediary nodes in the tree but can get the
grammar to look a lil ugly with so many `;'.