feat(PEG): Match structured data via mMatchList()
Now besides parsing arrays of chars into trees, this PEG
implementation is also able to match tree structures thanks to the
paper "Parsing Expression Grammars for Structured Data".

The function mMatchList started as a copy of the mMatch function and
does still have a lot of things that still need to be cleaned up.