feat(PEG.py): Generate CAPCHAR for rules used within capture blocks
The capture operator (%{}) should capture chars for everything that
matches within it. Even when the matches are declared in different
rules. Here's the smallest example: "S <- %{ A }; A <- 'a'"

A `CAPCHAR' instruction should be generated after the instruction
`CHAR a' in the above example. This is what this change implements.
The tree lookup that marks which terminal nodes should be captured to
satisfy the aforementioned case.