feat(PEG.py): Don't follow rules used outside block capture operators
The compiler processes the grammar tree and looks for block capture
operator %{} instances and then it tries to find all the identifiers
within its scope recursively within the grammar until it finds the
terminals to mark them to be captured. E.g.:

  Ident <- Space* %{ Prefix? IdentStart IdentCont* }
  IdentStart <- [a-zA-Z_]
  IdentCont <- IdentStart / [0-9]
  Prefix <- '$' Space*
  Space <- ' '

The above grammar allows matching the following inputs: "$var01, v,
v1, $V" etc. There's a special input that it might accept: "$
var". Not sure if this is a good usage example but it clearly shows
the effects of this change. The `Prefix' rule is included within the
block capture operator. And although `Prefix' references `Space'. The
compiler won't follow `Space' to generate a CAPCHAR for ` ' because
it's also being used outside the block capture operator. A message is
printed to the output mentioning the skipping.