Draft of new capture design, some random fixes & cleanup
This new capture design requires matches to be captured within the
stack frame.  When a given stack frame is popped from the stack within
the *Return* instruction, its captures are wrapped around a
`Value::Node` alongside with the name of the non-terminal.  This node
is then pushed to the accumulator register on the virtual machine.

The StackFrame struct was refactored and now there's a field for the
type of the frame (backtracking or function call) that is used within
the fail state to allow us to implement the bounding of the left
recursion.

This change doesn't really fully implement the capture, but it tests
the fundamental concept of matching functions setting the accumulator,
that is then collected by the `capture` instruction that should follow
both `Char` and `Call` instructions that are to have their matches
captured.