vm/parser: initial support for label based error reporting
This is not working properly within predicates yet as the flag
`within_predicate` still needs to be flipped then properly restored,
and I haven't found a nice way to do it yet.

Aside from the missing awareness on the predicate, this mostly
implements `throw.1` from "Syntax Error Recovery in Parsing Expression
Grammars" from S. Medeiros & F. Mascarenhas.  It also contains some of
the pieces for `throw.2` & `throw.3`, like the error log and the
recovery map.