parser: better resolving for deferred firsts/follows sets
A token that goes into the `firsts` set of is now first resolved
recursively before being passed into the final set of follows.  That
allows us to capture the follows set of sequence items that reference
recursive forward declarations.  The `firsts` set is also now being
deduped.

On top of the change above, this PR contains a bit of a cleanup in the
mix.  The final follows set is now computed as a step of the
backpatching process, which also became a thing in this commit,
instead of just being "the weird lines underneath the AST::Grammars
branch of Compiler.compile".