vm: fixes captures for left recursive grammars
The bounded left recursion implemented on this VM needs a way to
discard captured values of recursions that have been successful, but
weren't the longest match.

We had a draining mechanism for fail() already, and it was extended to
the return opcode as well.  The length of the match is saved when
increasing the bound, and the old values are discarded before the
capture vector is taken to be fed into the accumulator.