Overhaul how values are captured
There's now a `captures` vector and a `last_capture_committed` in both
the VM state and in each call frame.  The opcodes that implement
terminal matching will push values onto the captures vector of the
current call frame.

The return opcode pushes a new node with the `captures` vector of the
call frame that was popped from the stack.  If the call stack is
empty, values are pushed onto the virtual machine's `captures` vector.

For dealing with backtracking, the fail state will drain nodes from
`last_capture_committed` up to the length of the `captures` vector.
And the commit opcode will update the `last_capture_committed` index.

Unfortunately, this change breaks captures for left recursive call
frames.