vm/tests: a lot going on to try to get a few more LR tests to pass
These changes aren't by any means final.  But there are more
comprehensive tests for matching and capturing grammars with left
recursive calls.

The debugging in the VM is a bit of a mess right now, and needs to
have better compile time disabling alternatives.  But it has helped a
lot to indent the output with a width * call_stack.len().

The changes themselves to captures that helped making progress were:

 - last_capture_committed is now updated by the Choice opcode, and
   PartialCommit doesn't update it anymore;
 - the value returned by VM.run() is now the last one pushed into the
   capture stack, not the first;
 - lvar.4: wrapp all the values captured in the current frame with a
   `Node` named with the production's name.  It tries to flatten the
   recursive calls that ascend from alternatives deeper in the stack
   it also commits captures;
 - inc.3: like lvar.4, captures values in the stack into a `Node` and
   tries to flatten recursive calls as well;