vm: use CapStackFrame to track captures instead of the call frames
This is initially a refactoring that allows tracking captures
separately from call frames because the implementation looks more
consistent as there is no strange if checking if there are no call
frames to use the vector in the VM itself.

But this can also be used as a basis to try to push capture frames
around ordered choices specifically for handling the base case of
left recursive frames.