vm/compiler: fix left recursive captures with optional in base case
The idea of committing captures at `{lvar,inc}.1` stops working when
there is a Choice operation that can fail within the body of the
function that tries to increment the left recursive bound.  Because
the choice will also commit captures, we'll end up with duplicated
values within the stack at the base case.

This commit introduces three new instructions that take the main value
stack manipulation off the call/return/fail instructions and give it
to specific instructions that will push a new capture stack frame,
trigger a commit, or pop an item off the capture stack and push it
onto the current frame again.

And finally, we wrap the *Ordered Choice* operator with a pair of
Cap{Push,Pop} and add CapCommit instructions as the last instruction
of a branch of the Ordered Choice operator.