compiler: syntactically detect left recursion
We had a big and clumsy global flag that enabled or disabled
precedence for *all* the calls.  That is alright when it comes to
correctness, as the `lvar.*` rules are compatible with the `var.*`
ones, but for efficiency, especially when capturing values, that is a
big waste.

Up until this very commit, all the stuff the compiler did was in a
single pass.  The left recursion detection is the first sub-pass
that's introduced.  At some point it might make sense getting it done
in the single pass style, but my brain hurts just thinking about it.
This change makes the compiler slower for sure. But not by much I
suppose, and it gives the VM much more precise information about the
precedence.

The this left recursion detection pass is read-only.  It doesn't
transform the syntax tree in any, and the detector itself is a
recursive tree traversal that also contains a stack to detect mutual
recursion.