vm: change VM.run's input type to Value
This is another step towards the goal of matching data structures.
Instead of relying on generics or traits, we can rely on our own value
types as they're the ones that are both returned from the string to
list and received in list to list (or maybe one day list to output).

To keep a friendly interface available, the method `VM.run_str` was
added, so the string is mapped into a vector of `Value::Chr()`'s on
the user's behalf.