vm/parser: expand AST so it matches output of its matching
The newly added public function `expand` takes an AST (let's call it
original) and generates an AST that matches the output of a successful
parsing with the original AST.

The `vm` is mentioned here because there was a bug in the handler for
Instruction::Open.  It wasn't properly setting the cursor to zero to
allow matching the first item of the nested list.