vm: commit captures entering in a choice expr
This change fixes the error captured with the following test:

  let p = compile(&cc, "A <- '0x' [0-9a-fA-F]+ / '0'");
  assert_success("A[0xff]", run_str(&p, "0xff"));

The above assertion was failing, and the result looked like this:

  A[ff] (instead of A[0xff])

Upon closer inspection, one can see that the entire expression in A is
wrapped in a big Choice, and the first arm of this Choice call is a
single string match of two chars (`0x`) followed by a Choice operator
used within the One Or More expression, and another Choice operator
per entry in the Class expression.

Without commiting where the capture cursor is will cause the first
match (`0x`) to be completely dropped when the first Choice on '1-9'
fails.