vm: pass error label down to Error from return
The recovery node is now being built with both the label and the label
message, if one is attached.  The `recovery` field within `StackFrame`
was renamed to `recovery_label` and it's now an `Option<usize>` for
optionally carrying the recovery label that's being returned from.
The presence of a value here is also what signals return that it is
dealing with a recovery label.