lib: Add value type Node and use it as the default capture container
This change adds a data new data type to the virtual machine so that
we can better represent/store captured values.  Captures always come
from the Return state of the machine, so the name of the node is
always available.  This change also extend both parser and compiler to
recognize and emit a Close opcode marked with the proper container
type.  The formatters were also changed and the hacky special case for
lists was finally removed.