lang: Add @nocap decorator for suppressing captures locally
The general form of a decorator prefixes rule definitions and starts
with the at sign (`@') followed by an identifier.  Notice that no
spaces are allowed between the at sign and the identifier.

For supporting specifically the @nocap modifier, it was necessary to
add a new return opcode namely `ReturnNoCap' that signals the virtual
machine that the top of the capture stack should be discarded instead
of pushed onto its parent frame wrapped into a `Value::Node`.

A new field was added to the `ast::Definition' member, which led to
the changes in the compiler so it could emit the new opcode if it
found the `@nocap' decorator in a given `Definition'.  Both the hand
written parser and the `langlang.peg` grammars were updated to reflect
this change.