feat(lib/compiler): implement `unwrapped(P)` builtin
This checks statically if a semantic action expression is just a call
to `unwrapped(P)` and if it is, set the right capture type for the
RETURN in the expression used by the production.  The check prevents
calling `unwrapped()` from within a sub-expression, or calling it with
arity different from 1.