[WIP] Progress on semantic action parameters
The design here isn't clear yet, but there's a sense that if expected
signatures for each can be exhaustively inferred from bottom up, and
the input signature of the semantic action can be matched against the
output signature of the production.  This commit is rather unexciting
as it just saves the snapshot of whatever we got working.  But it even
leaves a gross problem open.  How to get rid of the parameters stuck
in the same stack frame as the output of a semantic action.