go: First bits of a Go code generator
The Value interface was introduced to be used as the output contract
of the *generated* parser.  e.g.: the handwritten parser that reads
the grammar input (grammar_parser.go) returns the Node interface,
which gets consumed by the code generator (gen_go.go).  The code
generator then emits a parser in which each parsing function will
return the new Value interface.

The difference between them is that the Node interface is specialized
to describe the AST nodes of a PEG gramamr.  And the Value interface
is meant to be generic to allow higher level translations to
specialize it into something actually useful to the user.