go: Add context to error reporting
This change introduces syntax for labeling expressions to the input
language and runtime support for the `Throw` operation in both grammar
parser and the Go generator.  The new syntax looks like the following:

   Production <- Expr^Label

An expression suffixed with a label (a labeled expression) will be
translated to `Expr / Throw(Label)`.  In other words, labeling an
expression doesn't change anything for expressions that succeed.  But
for expressions that fail, the error generated by `Throw` will not be
caught by the backtracking mechanism generating an error immediately.
The method `SetLabelMessages` was introduced to the `Parser`
interface, so users of the generated parser can set messages to be
used when a specific label is thrown.

There should be some documentation around how to use this operator, as
anything beyond the simple usage becomes too tricky too quickly.  The
very long story short though is that labeling should only be used on
expressions that can't fail at all.  e.g.: the parenthesis after an if
statement in Java Script.  If the parser consumes the `if` token,
it *must* see an opening parenthesis after, otherwise we know there's
a syntax error there.

Some very initial support for tracing the call graph was added through
the `Tracer` interface that's now embedded into the `Parser`
interface.