go[gen_go]: Allow setting output package and struct name
This change between the Go code generator and the CLI allows users to
customize both the name of the package (which defaults to `parser`)
and a suffix for the parser struct that will be output (defaults to no
suffix and just use the name `Parser`).  That allows generating
different parsers to live within the same Go package.

Here's an example:

  $ go run ./cmd/... -language go       \
      -go-package mygopkg               \
      -go-struct-suffix Stuff           \
      -grammar path/to/my/grammar.peg

The parser generated can be instantiated with:

  p:= NewParserStuff("input")