go/gen_go: Return nil if matching succeeds but doesn't consume input
Prior to this change, the output tree out return nodes with empty
expressions, which leads to the very confusing pattern of having to
check if a given node has data before actually using it.

To fix this bug from its inception, the solution proposed is to *not*
produce the output node if there isn't any subexpression to be
attached to it.  In a sub-expression, this works great because
`Sequence` was also changed to *not* include nil items.

The biggest, most apparent change, is when the top production doesn't
consume any input and still succeeds.  Leading to the need to check
for nil just once, right after calling the top parsing rule.  e.g.:

  p := parser.New(input)
  value, err := p.ParseRootExpr()
  if err != nil {
    return nil, err
  }
  if value == nil {
    return nil, nil
  }
  // this is now safe to use `value`