go/value: add Type() to the Value interface and improve text repr
With the Type() method, users of the Value() API have now a different
way of checking the type of a node besides using Go's type switch.  It
is a nice little shortcut more useful for debuggin than for any sort
of type checking, but hey it's good to have options.