go[gen_go]: draft of error recovery support
The code generator now gathers all labels and production names in
their respective sets, and when writing down the constructor, the new
field in the generated parser, recoveryTable, is populated associating
a label to a production function of the same name if it exists.  The
`Throw` operation in the generated parser will then first look up the
recoveryTable and if the label being thrown has a production
associated with it, that production is called and its return is
wrapped around a newly introduced Error value type.  Notice that throw
won't do anything if it is within a predicate now, it just throws a
regular error.