go: draft implementation of import directives
This change introduces new syntax and behavior to the input grammar
language.  It introduces import directives so rules of a grammar can
be used in other grammars.  The general syntax is as follows:

  @import Rule1, Rule2, RuleN... from "./path/to/dependency.peg"

The import is recursive, importing a file with other imports will also
work.  The path that can be used is very restricted right now.  Only
relative paths that explicitly start with "./" can be used.  This just
simplifies landing this feature and other import loaders can be
provided.