go[gen_go]: change the constructor's signature
This is a bit of a drastic change that removes the setting of the
input from the parser's constructor and leaves it to a second method
call to SetInput.

We get the ability to re-use the same parser struct between parsing of
different strings.  That is mostly because it wouldn't be ergonomical
to always have to set some state over and over again, like the mapping
between labels and error messages and the capturing of spaces.