rust: factor out ast and parser into new package langlang_syntax
This is not just a copy and paste, but rather a revamp with the main
goal of providing enough infrastructure to reimplement features
implemented exclusively in the Go side right now.

The new AST types are structs and there is one enum for the expression
types and another for literal types.  That allows storing metadata
like the span in the input in which a given AST node has been found.
There's now also a visitor that is rather straightforward for
traversing any level of the AST.

Notice that although the generated parsers aim to build CSTs, this
manual grammar parser doesn't generate a CST.  Testing the "roundtrip"
of parsing and then stringifying isn't possible with just one string,
we need both `input` and `expected` values in our tests.