rust: rewrite compiler to use new langlang_syntax package
The `Compiler::compile_node()` method was removed in favor of the
family of methods `Compiler::visit_*` implemented as part of the
Visitor trait that's now implemented for the `Compiler` struct.

Since the effort of changing the binary to also use the new package
was small enough, such change is also included in this commit.