rust: automated whitespace handling injection & lex operator runtime
These two things kinda have to be implemented together as the lex
operator is just a way of customizing how whitespace handling
injection works.  Or better, where it should not run.

The basic algorithm to add whitespace handling is straightforward: add
an Identifier("WS") before each item in a sequence of a non-syntactic
rule.  The devil is in the details though.  If the recursive traversal
of the rewrite finds a Choice operator that isn't syntactic, it will
return a Sequence with [Identifier("WS"), Choice(items)] in which
items is the list of choices that have had their expressions expanded
with the flag `consume_first` turned to false.  This transforms the
following expression:

  A <- B / C / D / "E"

Into:

  A <- WS (B / C / D / "E")

Which will correctly consume spaces before possibly hitting the
terminal match "e".

Note: this isn't tested at all with either lists or objects.  So
data-structure matching might not be working very well with this
feature for the time being.