rust: fix tests and revamp test structure
The `tests` module was moved off the `langlang_lib` crate and into the
`tests` directory within the root of the repository.  A new import
loader `InMemoryImportLoader` was added for particularly helping with
tests.  All these changes will allow porting the tests written in the
Golang prototype to their new home.