syntax/ast: rename node String_ to String
The advantages of having the underscore were both making sure it was
explicitly different from the standard library's String.  But there
are very few cases in which this node will be used without the ast
namespace anyway.  So the consistency might be worth the possible
annoyance of having to namespace the use of standard strings when
importing * from ast, which isn't dangerous and is self inflicted
anyway.