compiler: fix capturing 0-arity exprs at the cost of breaking leftrec
This is a complicated one; by creating a new capture frame in the
middle of an expression when we emit both ZeroOrMore and Option, the
popping of this frame overwrites the index of the committed capture
index of the frame left in the stack.

We can introduce a new opcode especially for 0-arity expressions, but
that's going to take a bit more time, so let's just fix this case and
let LRs break for now, as they're an extension and the core should be
prioritized over extensions when such dilemma arises.  Tests lr1 and
lr4 are breaking and so they're being marked as ignored on this
commit.