go: tiniest implementation of semantic actions
That shouldn't even be called semantic actions, as they are rather
limited.  The way they're implemented here allows users only to
transform Value nodes into other Value nodes.  These are going to be
mostly useful for making the CST less verbose before traversing it to
generate an AST with custom code.  That's the only case we currently
have in mind and this feature is most likely going to be deprecated
when there is a good design for more encompasing semantic actions.