rs/import: Labels must be serched as well for recovery rules
The dependency finder of a rule was alware of Identifiers used within
any subexpressions.  But we also need to check for labels which might
have recovery rules defined with the same name and add them as
dependencies as well.  e.g.:

  // File leaf.peg
  @import Val from "./dependecy.peg"
  Main <- Val ("," Val)*

  // File dependency.peg
  Val <- Hex / Dec / Str / Bool
  Hex <- '0x' [a-fA-F0-9]+^LabelHex
  LabelHex <- (!Val .)*

The importer wasn't pulling "LabelHex" into the file "leaf.peg"
without this change.

This also changes Comipler.compile to take the starting rule as an
optional value, and if it is not present, the first rule that appears
in the grammar ast will be used.