go[BREAKING CHANGE]: Overhaul naming of all the node types
This is in preparation for bootstrapping the go parser from the
grammar.

The first rename was in the `values.go` module.  The `Value` prefix
was stripped from {Value}String, {Value}Sequence, {Value}Node, and
{Value}Error.  The idea is to bolt this module into the final parser.

The above rename caused a collision between the `Node` interface in
the `grammar_ast.go` module.  Then this interface was renamed to
`AstNode`.

A mostly no-op change for now is that the prelude in go was split up
in its own file `prelude.go.in`.