go[BREAKING CHANGE]: Remove runtime dependencies from output parser
First and foremost, this is a breaking change because the output
parser name is now hard coded, and the constructor is called
`NewParser`.

Such change implies that from now on only a single parser is allowed
within a Go package, as the tool will not provide the necessary
options to avoid name collision.  The CLI parameter `go-prefix` was
removed in consequence of this change.

This decision simplifies the design of the overall design as well,
since now the `parser.go` file is both the base implementation of the
grammar parser, and the base implementation of the parser that is
generated, but the reuse is enabled during build time, instead of
requiring the output parser to depend on the tool as a library during
their runtime.

The Go generator now includes the files `parser.go`, `value.go`, and
`errors.go`.  Each of these files is parsed by `go/parser`, before
inclusion, both `package` and `import` statements are removed, and it
is the prelude that now injects these statements.  The package name is
still customizable through the `go-package` CLI option.

This change also opens up a clear path to bootstrapping most of the
`grammar_parser` automatically.  But that is a change for another day.

The generated parser also gets a pretty printer embedded into it.