go: low hanging fruit perf improvement on ChoiceRune
Before this change, ChoiceRune was unfolding into a regular Choice
with linear search through function calls and at least one more
allocation a the new slice of alternatives.

Now, the input to be attempted by ChoiceRune are passed as a map, and
on the happy path, there is now a map lookup replacing the linear
search, and no allocations at all.  If all alternatives fail, a slice
of the same length of the input map will be allocated as well as a new
error node taking the slice of attempted alternatives.

This is very low effort for a very good reward.  This is used by the
rule that parses spaces, which was glaringly showing in the CPU
profiler for even the simplest inputs.