go: Add label message to Error node
This is a breaking change, as we're adding a new parameter to the
constructor `NewError`.  That is useful when generating incremental
parsers.  That way, if the generated parser contains the proper labels
with recovery expressions associated with them, the Error node will
now also contain the error message associated through the
`SetLabelMessages` method.

Also added helper functions `writel(s)` and `writeil(s)` that print
out "\n" after printing string `s` and cleaned up all the go generator
using these helpers.