rs: Allow proper overriding of builtin imports
Ironically, the `langlang.peg` couldn't be parsed by itself because
the overriding of the `Spacing` rule needed to allow comments wasn't
working as expected.

First, we were unnecessarily injecting the parsed built-in stack frame
in every import level and now it was moved to after the dependency
check is finished.

And second, we were selectively picking some rules from the
`builtins.peg` file then picking their dependencies.  Now we're just
iterating through all the rules and adding them as the main grammar's
definitions right away (we assume builtins.peg depends on nothing).