go: New benchmark with langlang grammar
There's some performance work that needs to be done, and before
starting that work, we needed better tests.  Since the most
sophisticated grammar so far is langlang's, we're measuring how that
grammar is performing by running it against a few other grammars, then
against itself.

The existing benchmarks of previous versions were moved into a new
suite directory, and we now have the "import" and "langlang" suites.