go: memoize each and every production
PEGs are based on backtracking, so even when parsing succeeds, it is
highly likely that the parser stumbled upon quite a few different
errors and recovered from them (through the ORDERED CHOICE operator).

We can improve on this by caching the results of every production
that's executed by the parser.  This is what packrat parsers do to
guarantee that they finish on linear time.

There are much more sophisticated ways of caching AST nodes.  So this
solution isn't final, but it's a good self-contained improvement that
is very cheap to implement!