go: Get rid of fmt.Sprintf calls in the parser
Since the underlying algorithm of the parser relies so heavily on
generating errors and backtracking many times until it finds the right
path, any allocation on the hot path causes noticeable impact on the
overall allocation count.

This doesn't seem to improve simpler grammars, but it has shown much
better results on more complex grammars with complex inputs.  e.g.:
the langlang-langlang benchmark saw 34% of decrease in allocs.