go/gen_go: VisitSequenceNode: shorten path with no or single item
Since operations like Optional(?) and ZeroOrMore(*) can return nil,
sequences can be nil themselves.  In that case we skip this altogether
and write `nil` for the value `VisitSequenceNode` is supposed to
write.

If the sequence has just one item, the sequence code generation is
replaced by recursing into the only item in the list.

These two little checks will shrink the output of parsers that have
sequences, which are *very widely* used as they can be used when
dealing with operators that may not consume any input when succeeding.