go/ast: Pretty Printer for the Grammar AST
This commit extracts some of the tree printing present for the value
tree into its own module.  Then the generalized version is used by
both, the existing `ValuePrinter` as well as the new `grammarPrinter`.

For the `grammarPrinter`, a new method `PrettyPrint` was added to the
`AstNode` interface, so any AST node from the grammar can be pretty
printed with a single method call.