go: Revamp automatic space handling
The space handling logic was moved off of the `gen_go` file into a
discrete pre-processing step.  This makes it easier to implement
different generators, now that we don't have to re-implement injecting
productions related to space handling in each generator.

The method `Parser.SetCaptureSpaces` should be deprecated and we
should generalize a solution for opting into capturing and/or ignoring
certain node patterns.