go: Implementation of most of choice, and backtracking
An important tangential change here is that Grammar AST ChoiceNode is
now a binary operation (instead of Nary) and it is now right
associative.

On the Compiler side, it now emits code for the following operations:

 - Captures
 - Optional
 - Choice
 - Class (based on Choice)
 - Char (IString was removed for now)
 - Not now uses IChoicePred

Quite a few improvements were done to the visualization of a compiled
program as well.

On the vm, Choice and backtracking now work after a stack was
added (vm_stack.go).  And with that we're passing quite a few tests
now.