go: Fix `ZeroOrMore`, `Not` and `And` asm gen for O=0
The already compiler implements some stuff that "A Parsing Machine for
PEGs"[0], like `BackCommit` for `And`, `FailTwice` for `Not`, and
`PartialCommit` for `ZeroOrMore`.  And these are behind the flag
Compiler.Config.Optimize=1.  The tests are now running against both
Optimize settings (0 and 1), and on the setting =0, misplacement of
labels was causing a few failures fixed by this change.