go: Interpreter is aware of lines and columns
This shifts where backtrack frames are created back to the VM.
There's now a new invariant cost, we're checking if it's a '\n' every
time a new character is consumed.

The assertions in the VM tests were rewritten to expect the string
representation of the returned value nodes instead of nesting the
nodes manually.