go: First draft of nested captures
The `stack.capture()` function was changed to first try to push the
capture into the first `frameType_Capture` that appears when
traversing the stack from right to left.

And now, the method `VirtualMachine.newNode()` can check if the frame
has any already captured values, and if it does not, it will create a
`String` with the range between `frame.cursor` and `vm.cursor`.