go: Get other nested capture tests passing
This includes coming up with a likely temporary solution for
implicitly capturing string nodes within the scope of a pair of
instructions capture_{begin,end}.  This current solution is to scan
the captured nodes of a scope from left to right checking each node
for gaps on the left, then applying a final check for a possible gap
at the end (on the far right of the interval).

The reason why this solution is temporary is because this VM should
also handle expressions without any definitions, resembling more
Regular Expressions than Context Free Grammars.  And for that
functionality, there will be a need for explicitly attributing the
captures for specific parts of the matching expression.