go: Implement the first bits of the Throw mnemonic
This also fixes the size of the throw instruction, so it lets the VM
skip through the grammars that do have labeled failures and will
correctly signal that an error was found, although it still doesn't
have access to a table that's going to connect that labeled failure
with the label itself.

The interpreter also still doesn't have a recovery expression, so
throw currently fully interrupts the execution of the parsing program.