go[cosmetics]: Move the stack.capture function to the VirtualMachine
The `stack.values` arrangement didn't feel right, and it felt even
weirder to pass the vm into the stack.  To simplify it all, stack is
back to being a slice of frames (`[]frame`).

There is still `findCaptureFrame`, which seemed harmless to be kept
within the `stack` struct as it doesn't require the stack to be
different than a slice.  There is still the case to be made that it'd
be better to keep that outside the `stack` entirely when we start
trying to translate the go code into other languages.  But that's a
decision to be made later.