go[BREAKING CHANGE]: New Error Reporting
1. [BREAKING]: Likely, the biggest change here is the change in how
columns are indexed.  We're going from a zero-based to a one-based
index.  This is going to be more consistent with how lines are
index (they've been one-based since the beginning.)

2. [BREAKING]: Another noticeable change is in the message itself.  Two
main things changed there:

  * The new evaluator based parser keeps a set of all the terminals
  attempted (outside of predicates).  And it can be used to hit users
  what characters are actually allowed at that point.  If the VM can
  capture any expected characters, it will be part of the error
  message.

  * Another change in the error message itself was the display of what
  character is under FFP.

The new full error message for the TINY grammar and the following
example:

    123456789.....
    ______________
  1|n := 5;
  2|f := 1;
  3|repeat
  4|  f := f * n;
  5|  n := n - 1
  6|until (n < 1);
  7|write f;

As you can see, there is a missing semicolon at the end of line 5.
The error message produced is going to look like:

  Expected ';', '!', '=', '<', '>', '+', '-', '/', '*',  but got 'u' @ 6:1

3. [FIX]: The grammar compiler was miss-placing its choice label, which
was producing an invalid ASM.