go: Remove some duplication while injecting Spacing rules
The automatic space handling step of the grammar AST transformation
works by prefixing each node within a Sequence with a space.  This is
correct, however it's a tad bit wasteful when sequences are nested.
In that case, the nested sequence shouldn't need to add the first
space, and that's what this change is doing!